+++ src/main/kotlin/io/gladed/watchable/WatchableMap.kt
@@ -18,9 +18,8 @@ package io.gladed.watchable
 
 import kotlin.coroutines.CoroutineContext
 
-/** A [Watchable] wrapper for a [Map] which may also be modified or bound. */
-@UseExperimental(kotlinx.coroutines.ObsoleteCoroutinesApi::class,
-    kotlinx.coroutines.ExperimentalCoroutinesApi::class)
+/** A [Watchable] wrapper for a [Map] which may also be modified or bound. Use [watchableMapOf] to create.*/
+@UseExperimental(kotlinx.coroutines.ObsoleteCoroutinesApi::class, kotlinx.coroutines.ExperimentalCoroutinesApi::class)
 class WatchableMap<K, V> internal constructor(
     override val coroutineContext: CoroutineContext,
     initial: Map<K, V>
